Pager: Rewrite and introduce direct page links
This commit completely rewrites the Pager component to make better use
of horizontal space on desktop devices. We introduce direct, clickable
page numbers inbetween the "Previous" and "Next" link and automatically
shift them based on the currently viewed page. The component ensures
that the buttons do not shift, so the target click areas stay the same
regardless of the current page.

Mobile devices and small screens will continue to only see the
"Previous" and "Next" links.